feat(Overview): refactor column stats handling and improve click outside logic
