+++ shared/src/api/base/client.ts
@@ -83,9 +83,6 @@ const tagTypes = [
   'feedback',
   'folder',
   'folderColumnStats',
-  'taskColumnStats',
-  'productColumnStats',
-  'versionColumnStats',
   'hierarchy',
   'inbox',
   'info',
@@ -97,6 +94,7 @@ const tagTypes = [
   'overviewTask',
   'tasksFolder',
   'product',
+  'productColumnStats',
   'progress',
   'project',
   'projectAddons',
@@ -112,10 +110,12 @@ const tagTypes = [
   'siteSettingsSchema',
   'tag',
   'task',
+  'taskColumnStats',
   'team',
   'user',
   'userPool',
   'version',
+  'versionColumnStats',
   'view',
   'viewer',
   'watchers',
