+++ shared/src/api/queries/index.ts
@@ -5,6 +5,7 @@ export * from './addons'
 export * from './attributes'
 export * from './authentication'
 export * from './cloud'
+export * from './columnStats'
 export * from './config'
 export * from './entities'
 export * from './entityLists'
