+++ shared/src/api/queries/overview/getOverview.ts
@@ -3,24 +3,15 @@ import {
   GetTasksByParentQuery,
   GetTasksListQuery,
   GetFolderColumnStatsQuery,
-  GetFolderColumnStatsQueryVariables,
-  GetFolderColumnStatsDocument,
   GetTaskColumnStatsQuery,
-  GetTaskColumnStatsQueryVariables,
-  GetTaskColumnStatsDocument,
   foldersApi,
   SearchFoldersApiArg,
   GetTasksListQueryVariables,
 } from '@shared/api/generated'
 import { PubSub } from '@shared/util'
 import { EditorTaskNode } from '@shared/containers/ProjectTreeTable'
-import type { FieldStats } from '@shared/containers/ProjectTreeTable'
-// deep import: the barrel would create a runtime circular dependency (api ↔ containers)
-import {
-  normalizeFieldStats,
-  mergeFieldStats,
-} from '@shared/containers/ProjectTreeTable/components/TableFooter/mapColumnStats'
-import { hasNewTargetFields } from '@shared/containers/ProjectTreeTable/components/TableFooter/metricTargets'
+import type { FieldStats } from '../columnStats'
+import { normalizeFieldStats, mergeFieldStats, hasNewTargetFields } from '../columnStats'
 import {
   DefinitionsFromApi,
   FetchBaseQueryError,
@@ -121,6 +112,8 @@ type TagTypes = TagTypesFromApi<typeof gqlApi>
 type UpdatedDefinitions = Omit<Definitions, 'GetFilteredEntities'> & {
   GetTasksByParent: OverrideResultType<Definitions['GetTasksByParent'], EditorTaskNode[]>
   GetTasksList: OverrideResultType<Definitions['GetTasksList'], GetTasksListResult>
+  GetFolderColumnStats: OverrideResultType<Definitions['GetFolderColumnStats'], FieldStats[]>
+  GetTaskColumnStats: OverrideResultType<Definitions['GetTaskColumnStats'], FieldStats[]>
 }
 
 // GRAPHQL API
@@ -141,6 +134,24 @@ const enhancedApi = gqlApi.enhanceEndpoints<TagTypes, UpdatedDefinitions>({
       }),
       providesTags: (result, _e, { projectName }) =>
         getOverviewTaskTags(result?.tasks || [], projectName),
+    },
+    // footer stats: `targets` excluded from cache key + responses merged,
+    // so column toggles reuse cache and only added targets refetch
+    GetFolderColumnStats: {
+      transformResponse: (res: GetFolderColumnStatsQuery) =>
+        normalizeFieldStats(res?.project?.folders?.fieldStats ?? []),
+      serializeQueryArgs: ({ queryArgs: { targets: _t, ...rest } }) => rest,
+      merge: (cache, incoming) => mergeFieldStats(incoming, cache),
+      forceRefetch: ({ currentArg, previousArg }) => hasNewTargetFields(currentArg, previousArg),
+      providesTags: (_r, _e, { projectName }) => [{ type: 'folderColumnStats', id: projectName }],
+    },
+    GetTaskColumnStats: {
+      transformResponse: (res: GetTaskColumnStatsQuery) =>
+        normalizeFieldStats(res?.project?.tasks?.fieldStats ?? []),
+      serializeQueryArgs: ({ queryArgs: { targets: _t, ...rest } }) => rest,
+      merge: (cache, incoming) => mergeFieldStats(incoming, cache),
+      forceRefetch: ({ currentArg, previousArg }) => hasNewTargetFields(currentArg, previousArg),
+      providesTags: (_r, _e, { projectName }) => [{ type: 'taskColumnStats', id: projectName }],
     },
   },
 })
@@ -618,26 +629,6 @@ const injectedApi = enhancedApi.injectEndpoints({
       },
       providesTags: (result, _e, { projectName }) =>
         getOverviewTaskTags(result?.tasks, projectName),
-    }),
-    // footer stats: `targets` excluded from cache key + responses merged,
-    // so column toggles reuse cache and only added targets refetch
-    getFolderColumnStats: build.query<FieldStats[], GetFolderColumnStatsQueryVariables>({
-      query: (variables) => ({ document: GetFolderColumnStatsDocument.toString(), variables }),
-      transformResponse: (res: GetFolderColumnStatsQuery) =>
-        normalizeFieldStats(res?.project?.folders?.fieldStats ?? []),
-      serializeQueryArgs: ({ queryArgs: { targets: _t, ...rest } }) => rest,
-      merge: (cache, incoming) => mergeFieldStats(incoming, cache),
-      forceRefetch: ({ currentArg, previousArg }) => hasNewTargetFields(currentArg, previousArg),
-      providesTags: (_r, _e, { projectName }) => [{ type: 'folderColumnStats', id: projectName }],
-    }),
-    getTaskColumnStats: build.query<FieldStats[], GetTaskColumnStatsQueryVariables>({
-      query: (variables) => ({ document: GetTaskColumnStatsDocument.toString(), variables }),
-      transformResponse: (res: GetTaskColumnStatsQuery) =>
-        normalizeFieldStats(res?.project?.tasks?.fieldStats ?? []),
-      serializeQueryArgs: ({ queryArgs: { targets: _t, ...rest } }) => rest,
-      merge: (cache, incoming) => mergeFieldStats(incoming, cache),
-      forceRefetch: ({ currentArg, previousArg }) => hasNewTargetFields(currentArg, previousArg),
-      providesTags: (_r, _e, { projectName }) => [{ type: 'taskColumnStats', id: projectName }],
     }),
   }),
 })
