+++ shared/src/api/queries/versions/getVersionsProducts.ts
@@ -24,12 +24,8 @@ import {
   GetProductsQuery,
   GetProductsQueryVariables,
   GetProductsColumnStatsQuery,
-  GetProductsColumnStatsQueryVariables,
-  GetProductsColumnStatsDocument,
   GetVersionsByProductIdQueryVariables,
   GetVersionsColumnStatsQuery,
-  GetVersionsColumnStatsQueryVariables,
-  GetVersionsColumnStatsDocument,
   GetVersionsQuery,
   GetVersionsQueryVariables,
   gqlApi,
@@ -46,12 +42,8 @@ import {
   transformVersionsResponse,
 } from './getVersionsProductsUtils'
 import { PubSub } from '@shared/util'
-import type { FieldStats } from '@shared/containers/ProjectTreeTable'
-import {
-  normalizeFieldStats,
-  mergeFieldStats,
-} from '@shared/containers/ProjectTreeTable/components/TableFooter/mapColumnStats'
-import { hasNewTargetFields } from '@shared/containers/ProjectTreeTable/components/TableFooter/metricTargets'
+import type { FieldStats } from '../columnStats'
+import { normalizeFieldStats, mergeFieldStats, hasNewTargetFields } from '../columnStats'
 
 // SHARED CACHE UPDATE HELPERS
 // These helpers are used by PubSub handlers to update cached data in real-time
@@ -236,6 +228,8 @@ type UpdatedDefinitions = Omit<Definitions, 'GetVersions'> & {
   GetVersions: OverrideResultType<Definitions['GetVersions'], GetVersionsResult>
   GetVersionsByProductId: OverrideResultType<Definitions['GetVersions'], GetVersionsResult>
   GetProducts: OverrideResultType<Definitions['GetProducts'], GetProductsResult>
+  GetProductsColumnStats: OverrideResultType<Definitions['GetProductsColumnStats'], FieldStats[]>
+  GetVersionsColumnStats: OverrideResultType<Definitions['GetVersionsColumnStats'], FieldStats[]>
 }
 
 const enhancedVersionsPageApi = gqlApi.enhanceEndpoints<TagTypes, UpdatedDefinitions>({
@@ -254,6 +248,24 @@ const enhancedVersionsPageApi = gqlApi.enhanceEndpoints<TagTypes, UpdatedDefinit
     GetProducts: {
       transformResponse: transformProductsResponse,
       providesTags: provideTagsForProductsResult,
+    },
+    // footer stats: `targets` excluded from cache key + responses merged,
+    // so column toggles reuse cache and only added targets refetch
+    GetProductsColumnStats: {
+      transformResponse: (res: GetProductsColumnStatsQuery) =>
+        normalizeFieldStats(res?.project?.products?.fieldStats ?? []),
+      serializeQueryArgs: ({ queryArgs: { targets: _t, ...rest } }) => rest,
+      merge: (cache, incoming) => mergeFieldStats(incoming, cache),
+      forceRefetch: ({ currentArg, previousArg }) => hasNewTargetFields(currentArg, previousArg),
+      providesTags: (_r, _e, { projectName }) => [{ type: 'productColumnStats', id: projectName }],
+    },
+    GetVersionsColumnStats: {
+      transformResponse: (res: GetVersionsColumnStatsQuery) =>
+        normalizeFieldStats(res?.project?.versions?.fieldStats ?? []),
+      serializeQueryArgs: ({ queryArgs: { targets: _t, ...rest } }) => rest,
+      merge: (cache, incoming) => mergeFieldStats(incoming, cache),
+      forceRefetch: ({ currentArg, previousArg }) => hasNewTargetFields(currentArg, previousArg),
+      providesTags: (_r, _e, { projectName }) => [{ type: 'versionColumnStats', id: projectName }],
     },
   },
 })
@@ -929,33 +941,12 @@ const injectedVersionsPageApi = enhancedVersionsPageApi.injectEndpoints({
   }),
 })
 
-// footer stats: `targets` excluded from cache key + responses merged,
-// so column toggles reuse cache and only added targets refetch
-const vpStatsApi = injectedVersionsPageApi.injectEndpoints({
-  endpoints: (build) => ({
-    getProductsColumnStats: build.query<FieldStats[], GetProductsColumnStatsQueryVariables>({
-      query: (variables) => ({ document: GetProductsColumnStatsDocument.toString(), variables }),
-      transformResponse: (res: GetProductsColumnStatsQuery) =>
-        normalizeFieldStats(res?.project?.products?.fieldStats ?? []),
-      serializeQueryArgs: ({ queryArgs: { targets: _t, ...rest } }) => rest,
-      merge: (cache, incoming) => mergeFieldStats(incoming, cache),
-      forceRefetch: ({ currentArg, previousArg }) => hasNewTargetFields(currentArg, previousArg),
-      providesTags: (_r, _e, { projectName }) => [{ type: 'productColumnStats', id: projectName }],
-    }),
-    getVersionsColumnStats: build.query<FieldStats[], GetVersionsColumnStatsQueryVariables>({
-      query: (variables) => ({ document: GetVersionsColumnStatsDocument.toString(), variables }),
-      transformResponse: (res: GetVersionsColumnStatsQuery) =>
-        normalizeFieldStats(res?.project?.versions?.fieldStats ?? []),
-      serializeQueryArgs: ({ queryArgs: { targets: _t, ...rest } }) => rest,
-      merge: (cache, incoming) => mergeFieldStats(incoming, cache),
-      forceRefetch: ({ currentArg, previousArg }) => hasNewTargetFields(currentArg, previousArg),
-      providesTags: (_r, _e, { projectName }) => [{ type: 'versionColumnStats', id: projectName }],
-    }),
-  }),
-})
-
 // export gql endpoints
-export const { useGetVersionsQuery } = enhancedVersionsPageApi
+export const {
+  useGetVersionsQuery,
+  useGetProductsColumnStatsQuery,
+  useGetVersionsColumnStatsQuery,
+} = enhancedVersionsPageApi
 // export custom queries
 export const {
   useGetVersionsInfiniteInfiniteQuery: useGetVersionsInfiniteQuery,
@@ -963,7 +954,6 @@ export const {
   useGetProductsInfiniteInfiniteQuery: useGetProductsInfiniteQuery,
   useGetGroupedVersionsListQuery,
 } = injectedVersionsPageApi
-export const { useGetProductsColumnStatsQuery, useGetVersionsColumnStatsQuery } = vpStatsApi
 
 // export API instances for cache manipulation
 export { enhancedVersionsPageApi, injectedVersionsPageApi }
