+++ shared/src/containers/ProjectTreeTable/components/TableFooter/mockFieldStats.ts
@@ -1,13 +1,11 @@
 import { FieldStats } from './mapColumnStats'
 
 export const mockFieldStats: FieldStats[] = [
-
   {
     columnName: 'name',
-    folderCount: 1599,
-    taskCount: 8423,
+    primaryCount: 1599,
+    secondaryCount: 8423,
   },
-
   {
     columnName: 'attrib_fps',
     sum: 39975,
@@ -16,7 +14,6 @@ export const mockFieldStats: FieldStats[] = [
     columnName: 'attrib_frameStart',
     sum: 1601000,
   },
-
   {
     columnName: 'createdAt',
     min: '2026-01-12T09:30:00Z',
